Guard headers against multiple includes
M2 added optional support for `#include` about 4 months ago,
however it didn't detect when it's including the same header
multiple times, and would error when attempting to add the
duplicate definitions. This should fix that error, and as
a bonus, let users performing the bootstrap process to not
list every M2libc dependency when compiling kaem and hex2.